[scene] - Refactor out into scene loader abstraction and gltfLoader. Added VulkanSDK 1.0.30.1
