+++ TLVulkanRenderer/src/scene/Scene.h
@@ -1,13 +1,12 @@
 #pragma once
 
-#include <map>
-#include "tinygltfloader/tiny_gltf_loader.h"
 #include "SceneUtil.h"
 #include "scene/Camera.h"
 #include <geometry/Geometry.h>
 #include <geometry/materials/LambertMaterial.h>
 #include <geometry/SBVH.h>
 #include "lights/Light.h"
+#include "sceneLoaders/SceneLoader.h"
 
 
 class Scene {
@@ -15,6 +14,7 @@ public:
 	Scene(std::string fileName, std::map<std::string, std::string>& config);
 	~Scene();
 
+	void ParseSceneFile(std::string fileName);
 	Intersection GetIntersection(const Ray& ray);
 	bool DoesIntersect(const Ray& ray);
 
@@ -33,8 +33,12 @@ public:
 	std::vector<Light*> lights;
 	SBVH m_sbvh;
 
+
 private:
-	
+
+	void PrepareTestScene();
+
+	std::unique_ptr<SceneLoader> m_sceneLoader;	
 	bool m_useSBVH;
 
 };
