+++ TLVulkanRenderer/src/scene/SceneUtil.h
@@ -2,6 +2,7 @@
 #include "Typedef.h"
 #include <glm/glm.hpp>
 #include <map>
+#include <vector>
 
 // ---------
 // VERTEX
