+++ TLVulkanRenderer/src/scene/sceneLoaders/SceneLoader.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class Scene;
+
+class SceneLoader
+{
+public:
+	virtual bool Load(std::string fileName, Scene* scene) = 0;
+};
