+++ TLVulkanRenderer/src/scene/sceneLoaders/gltfLoader.h
@@ -1,0 +1,10 @@
+#pragma once
+#include <map>
+#include "tinygltfloader/tiny_gltf_loader.h"
+#include "SceneLoader.h"
+
+class gltfLoader : public SceneLoader
+{
+public:
+	bool Load(std::string fileName, Scene* scene) override;
+};
